【第21节】 1. 充值仍有部分问题 2. 增加了 awardQuery，专门处理对API对奖品的查询

+++ big-market-domain/src/main/java/com/dasi/domain/strategy/service/rule/tree/impl/RuleLockTree.java
@@ -21,8 +21,8 @@ public class RuleLockTree implements IRuleTree {
 
     @Override
     public RuleCheckResult logic(String userId, Long strategyId, Integer awardId, String ruleValue) {
-        long limitLotteryCount = Long.parseLong(ruleValue);
-        long userLotteryCount = strategyRepository.queryUserLotteryCount(userId, strategyId);
+        int limitLotteryCount = Integer.parseInt(ruleValue);
+        int userLotteryCount = strategyRepository.queryUserLotteryCount(userId, strategyId);
         if (userLotteryCount > limitLotteryCount) {
             log.info("【策略规则树】rule_lock 放行：userLotteryCount={}, limitLotteryCount={}", userLotteryCount, limitLotteryCount);
             return RuleCheckResult.builder()
